+++ src/main/java/org/synyx/urlaubsverwaltung/restapi/absence/AbsenceController.java
@@ -5,8 +5,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.joda.time.DateMidnight;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.synyx.urlaubsverwaltung.core.application.domain.Application;
@@ -51,7 +51,7 @@ public class AbsenceController {
         value = "Get all absences for a certain period and person",
         notes = "Get all absences for a certain period and person"
     )
-    @RequestMapping(value = "/absences", method = RequestMethod.GET)
+    @GetMapping("/absences")
     public ResponseWrapper<DayAbsenceList> personsVacations(
         @ApiParam(value = "Year to get the absences for", defaultValue = RestApiDateFormat.EXAMPLE_YEAR)
         @RequestParam("year")