+++ src/main/java/org/synyx/urlaubsverwaltung/restapi/availability/AvailabilityController.java
@@ -5,8 +5,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.joda.time.DateMidnight;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.synyx.urlaubsverwaltung.core.person.Person;
@@ -40,7 +40,7 @@ public class AvailabilityController {
         notes =
             "Get all availabilities for a certain period and person. Maximum allowed period per request is one month."
     )
-    @RequestMapping(value = "/availabilities", method = RequestMethod.GET)
+    @GetMapping("/availabilities")
     public AvailabilityList personsAvailabilities(
         @ApiParam(value = "start of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_FIRST_DAY_OF_YEAR)
         @RequestParam("from")